test_taas.py: Move to unittest framework
This migrates our existing tests to the the unittest framework.
Specifically:

- Tests are now methods in `test_taas.py`.
- `test_process_csv` is no longer in a separate file.
